Introduce Backend::Api::Sources::File
Move file related methods from `Backend::Api::Sources::Package` into its
own class.
