+++ src/api/app/lib/backend/api/sources/package.rb
@@ -134,24 +134,6 @@ module Backend
                                                        params: options.compact, accepted: %i[repository arch])
         end
 
-        # Returns the content of the source file
-        # @return [String]
-        def self.file(project_name, package_name, file_name)
-          http_get(['/source/:project/:package/:filename', project_name, package_name, file_name])
-        end
-
-        # Returns the content of the source file
-        # @return [String]
-        def self.blame(project_name, package_name, file_name, options = {})
-          http_get(['/source/:project/:package/:filename', project_name, package_name, file_name], defaults: { view: :blame }, params: options, accepted: %i[meta deleted expand rev view])
-        end
-
-        # Writes the content of the source file
-        # @return [String]
-        def self.write_file(project_name, package_name, file_name, content = '', params = {})
-          http_put(['/source/:project/:package/:filename', project_name, package_name, file_name], data: content, params: params)
-        end
-
         # Writes source filelist to the package
         # @return [String]
         def self.write_filelist(project_name, package_name, filelist, params = {})
@@ -167,11 +149,6 @@ module Backend
         def self.undelete(project_name, package_name, options = {})
           http_post(['/source/:project/:package', project_name, package_name], defaults: { cmd: :undelete },
                                                                                params: options, accepted: %i[user comment time])
-        end
-
-        # Deletes a package source file
-        def self.delete_file(project_name, package_name, filename)
-          http_delete(['/source/:project/:package/:filename', project_name, package_name, filename])
         end
       end
     end
