+++ src/api/app/models/service.rb
@@ -85,13 +85,13 @@ class Service
 
     if document.xpath('//services/service').empty?
       begin
-        Backend::Api::Sources::Package.delete_file(project.name, package.name, '_service')
+        Backend::Api::Sources::File.delete(project.name, package.name, '_service')
       rescue Backend::NotFoundError
         # to be ignored, if it's gone, it's gone
       end
     else
-      Backend::Api::Sources::Package.write_file(project.name, package.name, '_service', document.root.to_xml,
-                                                comment: 'Modified via webui', user: User.session!.login)
+      Backend::Api::Sources::File.write(project.name, package.name, '_service', document.root.to_xml,
+                                        comment: 'Modified via webui', user: User.session!.login)
       service_package = Package.get_by_project_and_name(project.name, package.name, follow_project_links: false)
       return false unless User.session!.can_modify?(service_package)
